# TIG-1098 Add standalone benchmark binary to measure genny perf impact