--- a/src/canaries/include/canaries/Loops.hpp
+++ b/src/canaries/include/canaries/Loops.hpp
@@ -15,11 +15,8 @@
 #ifndef HEADER_C8D12C92_7C08_4B7F_9857_8B25F7258BB9_INCLUDED
 #define HEADER_C8D12C92_7C08_4B7F_9857_8B25F7258BB9_INCLUDED
 
-#include <iostream>
 
-#include <gennylib/PhaseLoop.hpp>
-#include <gennylib/context.hpp>
-
+#include <cstdint>
 
 namespace genny::canaries {