Most of what's described in TIG-1098 is covered. `db.ping()` is missing, but I'm planning to do it as part of TIG-1337 now that the ticket has been reopened. I'll probably also file a ticket to integrate the canaries with DSI. 

Right now the output is in a human readable format:

```bash
➜  genny git:(tig1098) ./dist/bin/genny-canaries "simple-loop" "phase-loop" "metrics-loop" "real-loop" -p false
simple-loop:	2143012ns
phase-loop:	50859249ns
metrics-loop:	192317833ns
real-loop:	219668218ns
➜  genny git:(tig1098) ./dist/bin/genny-canaries "simple-loop" "phase-loop" "metrics-loop" "real-loop" -p false
simple-loop:	2108709ns
phase-loop:	51095903ns
metrics-loop:	194208899ns
real-loop:	220928299ns
➜  genny git:(tig1098) ./dist/bin/genny-canaries "simple-loop" "phase-loop" "metrics-loop" "real-loop" -p false
simple-loop:	2282825ns
phase-loop:	51424344ns
metrics-loop:	192888125ns
real-loop:	219086235ns
```