@guoyr 2019-04-17
> This looks good but how do you envision this being used?

Run them as part of all genny DSI tasks? I think having these output to the same metrics format as genny so the metrics parsing scripts can be reused.

> Does it become another test type in dsi?

Hopefully the existing `GennyRunner` in DSI can be modified to run the canaries as well.