@guoyr 2019-04-17
I filed https://jira.mongodb.org/browse/TIG-1614 for followup work to 

1. report the canaries in Genny's csv2 format
2. integrate with DSI.

Let me know if you have concerns or other ideas.